+++ next/web/src/App/Admin/Stats/utils.tsx
@@ -5,42 +5,6 @@ import _ from 'lodash';
 import { relativeDateGetters } from '@/utils/date-range';
 import { useSearchParams } from '@/utils/useSearchParams';
 
-export const STATS_FIELD = [
-  'created',
-  'closed',
-  'reopened',
-  'conversion',
-  // 'internalConversion',
-  // 'externalConversion',
-  'firstReplyTimeAVG',
-  'replyTimeAVG',
-  'naturalReplyTimeAVG',
-  'replyCount',
-  'internalReplyCount',
-] as const;
-export type StatsField = typeof STATS_FIELD[number];
-export const STATS_FIELD_LOCALE: Record<string, string> = {
-  created: '新建工单',
-  closed: '关单数',
-  reopened: '激活工单数',
-  conversion: '流转数',
-  // internalConversion: '内部流转数',
-  // externalConversion: '外部流转数',
-  firstReplyTimeAVG: '平均首次回复时间',
-  replyTimeAVG: '平均回复时间',
-  naturalReplyTimeAVG: '平均回复自然时间',
-  replyCount: '对外回复数',
-  internalReplyCount: '对内回复数',
-};
-export const STATUS_LOCALE: Record<string, string> = {
-  notProcessed: '未处理',
-  waitingCustomer: '等待用户回复',
-  waitingCustomerService: '等待客服回复',
-  preFulfilled: '待用户确认解决',
-  fulfilled: '已解决',
-  closed: '已关闭',
-};
-
 const RANGE_DATE = ['lastSevenDays', 'lastWeek', 'week', 'month', 'lastMonth'] as const;
 const RANGE_DATE_LOCALE = {
   lastSevenDays: '最近 7 天',
@@ -49,8 +13,10 @@ const RANGE_DATE_LOCALE = {
   month: '本月',
   lastMonth: '上个月',
 };
-const defaultDateRange = relativeDateGetters['lastSevenDays']();
-export const useRangePicker = (fmt = 'YYYY-MM-DD') => {
+export const useRangePicker = (
+  fmt = 'YYYY-MM-DD',
+  defaultDateRange = relativeDateGetters['lastSevenDays']()
+) => {
   const [{ from, to, ...rest }, { set }] = useSearchParams();
   const rangeDates = useMemo(() => {
     return RANGE_DATE.reduce(
